Add files via upload
This version now uses millis instaed of waiting for the first two gates to finish closing, so that a train entering from the opposite direction does not get missed, and the gates will not close at an unwanted time.  The LDR detectors are also read twice and averaged to prevent spurious triggering due to electrical noise - mainly the DCC bus / track near the sensors.